+++ app/src/main/java/org/simple/clinic/teleconsultlog/prescription/medicines/TeleconsultMedicinesUpdate.kt
@@ -3,6 +3,9 @@ package org.simple.clinic.teleconsultlog.prescription.medicines
 import com.spotify.mobius.Next
 import com.spotify.mobius.Next.next
 import com.spotify.mobius.Update
+import org.simple.clinic.drugs.OpenIntention
+import org.simple.clinic.drugs.OpenIntention.AddNewMedicine
+import org.simple.clinic.drugs.OpenIntention.RefillMedicine
 import org.simple.clinic.mobius.dispatch
 
 class TeleconsultMedicinesUpdate : Update<TeleconsultMedicinesModel, TeleconsultMedicinesEvent, TeleconsultMedicinesEffect> {
@@ -13,11 +16,22 @@ class TeleconsultMedicinesUpdate : Update<TeleconsultMedicinesModel, Teleconsult
   ): Next<TeleconsultMedicinesModel, TeleconsultMedicinesEffect> {
     return when (event) {
       is PatientMedicinesLoaded -> next(model.medicinesLoaded(event.medicines))
-      EditMedicinesClicked -> dispatch(OpenEditMedicines(model.patientUuid))
+      EditMedicinesClicked -> {
+        val openIntention = openIntentionForEditMedicineScreen(model)
+        dispatch(OpenEditMedicines(model.patientUuid, openIntention))
+      }
       is DrugDurationClicked -> dispatch(OpenDrugDurationSheet(event.prescription))
       is DrugFrequencyClicked -> dispatch(OpenDrugFrequencySheet(event.prescription))
       is DrugDurationChanged -> dispatch(UpdateDrugDuration(event.prescriptionUuid, event.duration))
       is DrugFrequencyChanged -> dispatch(UpdateDrugFrequency(event.prescriptionUuid, event.frequency))
+    }
+  }
+
+  private fun openIntentionForEditMedicineScreen(model: TeleconsultMedicinesModel): OpenIntention {
+    return if (model.medicinesNotNullorEmpty) {
+      RefillMedicine
+    } else {
+      AddNewMedicine
     }
   }
 }